+++ src/main/java/com/zea7ot/leetcode/logs.txt
@@ -14,8 +14,8 @@
 5: 
 
 20201003:
-3: 0452, 0840, 1138
-4: 
+3: 0452, 0583, 0712, 0840, 1138, 1143
+4: 0072
 5: 
 
 20201002:
