+++ src/main/java/com/zea7ot/leetcode/tags/by_data_structure/string/by_topic/char_operation.txt
@@ -21,6 +21,8 @@ References:
 
 # Delete
 
+3   0583    https://leetcode.com/problems/delete-operation-for-two-strings/
+
 4   0072    https://leetcode.com/problems/edit-distance/
