+++ src/main/java/com/zea7ot/leetcode/tags/by_data_structure/string/by_topic/edit_distance.txt
@@ -7,4 +7,6 @@ References:
 ## By Topic
 # Edit Distance
 
+3   0583    https://leetcode.com/problems/delete-operation-for-two-strings/
+
 4   0072    https://leetcode.com/problems/edit-distance/
